+++ src/app-config/dicts.ts
@@ -252,7 +252,7 @@ export function getALlDicts () {
        * the preferred height is used and a mask with a view-more button is shown.
        * Otherwise the content height is used.
        */
-      preferredHeight: 210,
+      preferredHeight: 320,
       /** Word count to start searching */
       selectionWC: {
         min: 1,
@@ -563,7 +563,7 @@ export function getALlDicts () {
        * the preferred height is used and a mask with a view-more button is shown.
        * Otherwise the content height is used.
        */
-      preferredHeight: 210,
+      preferredHeight: 320,
       /** Word count to start searching */
       selectionWC: {
         min: 1,