+++ src/app-config/merge-config.ts
@@ -55,12 +55,22 @@ export function mergeConfig (oldConfig: AppConfig, baseConfig?: AppConfig): AppC
   merge('panelMode.instant.key', val => val === 'direct' || val === 'ctrl' || val === 'alt')
   mergeNumber('panelMode.instant.delay')
 
+  mergeBoolean('qsPanelMode.direct')
+  mergeBoolean('qsPanelMode.double')
+  mergeBoolean('qsPanelMode.ctrl')
+  mergeBoolean('qsPanelMode.instant.enable')
+  merge('qsPanelMode.instant.key', val => val === 'direct' || val === 'ctrl' || val === 'alt')
+  mergeNumber('qsPanelMode.instant.delay')
+
   mergeNumber('doubleClickDelay')
 
   mergeBoolean('tripleCtrl')
   merge('tripleCtrlPreload', val => val === '' || val === 'clipboard' || val === 'selection')
   mergeBoolean('tripleCtrlAuto')
   merge('tripleCtrlLocation', val => val >= 0 && val <= 8)
+  mergeBoolean('tripleCtrlStandalone')
+  mergeNumber('tripleCtrlHeight')
+  mergeBoolean('tripleCtrlPageSel')
 
   merge('baPreload', val => val === '' || val === 'clipboard' || val === 'selection')
   mergeBoolean('baAuto')