+++ src/background/initialization.ts
@@ -3,6 +3,7 @@ import checkUpdate from '@/_helpers/check-update'
 import { AppConfigMutable } from '@/app-config'
 import { getActiveConfig, updateActiveConfig, initConfig } from '@/_helpers/config-manager'
 import { init as initMenus, openPDF, openGoogle, openYoudao } from './context-menus'
+import { openQSPanel } from './server'
 import { init as initPdf } from './pdf-sniffer'
 import { MsgType, MsgQueryPanelState } from '@/typings/message'
 
@@ -49,6 +50,9 @@ browser.commands.onCommand.addListener(command => {
           updateActiveConfig(config)
         })
       })
+      break
+    case 'open-quick-search':
+      openQSPanel()
       break
     case 'open-google':
       openGoogle()
@@ -102,6 +106,20 @@ function onStartup (): void {
         })
       }
     })
+
+  // Chrome fails to inject css via manifest if the page is loaded
+  // as "last opened tabs" when browser opens.
+  setTimeout(() => {
+    browser.tabs.query({})
+      .then(tabs => {
+        tabs.forEach(({ id, url }) => {
+          if (id && url && url.startsWith('http')) {
+            browser.tabs.insertCSS(id, { file: '/content.css' }).catch(() => {/* noop */})
+            browser.tabs.executeScript(id, { file: '/content.js' }).catch(() => {/* noop */})
+          }
+        })
+      })
+  }, 1000)
 }
 
 function genClickListener (url: string) {