+++ src/typings/global.d.ts
@@ -1,4 +1,6 @@
 interface Window {
+  __SALADICT_PANEL_LOADED__?: boolean
+
   // For self page messaging
   pageId?: number | string
   faviconURL?: string
@@ -8,6 +10,7 @@ interface Window {
   __SALADICT_INTERNAL_PAGE__?: boolean
   __SALADICT_OPTIONS_PAGE__?: boolean
   __SALADICT_POPUP_PAGE__?: boolean
+  __SALADICT_QUICK_SEARCH_PAGE__?: boolean
 
   // Options page
   __SALADICT_LAST_SEARCH__?: string